Add a new Interface "getAccountsCustomersBoxs"
Modify "ReportLossService"

+++ safe_demo/src/main/java/com/temp/dao/LogDao.java
@@ -24,5 +24,5 @@ public interface LogDao {
 	 * @return boolean
 	 */
 	boolean setReportLossLog(ReportLossAction reportLossAction, 
-			int boxId, int reportLossType, int paymentType, float feeTotal);
+			long boxId, int reportLossType, int paymentType, float feeTotal);
 }
